third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/xmlhttprequest-response-type-blob-expected.txt

Request a file as the Blob type with asynchronous XMLHttpRequest in a document context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS req.status is 200
PASS req.response.size is 4
PASS buffer is "PASS"
PASS Set responseType before open(): Successful
PASS req.status is 200
PASS req.response.size is 4
PASS buffer is "PASS"
PASS Set responseType before send(): Successful
PASS req.status is 200
PASS req.response.size is 4
PASS buffer is "PASS"
PASS Set responseType in HEADERS_RECEIVED: Successful
PASS successfullyParsed is true

TEST COMPLETE
